next step - use customHooks
